device/requirements.txt
aioserial==1.3.0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
idna==3.1
influxdb-client==1.15.0
multidict==5.1.0
pyserial==3.5
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2021.1
Rx==3.1.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.4
yarl==1.6.3
